Features added. 1. swipe left to delete history 2. undo option to prevent a history delete

+++ app/src/main/java/com/gmail/monajemi/am/android/qrscanner/fragment/HistoryFragment.java
@@ -1,5 +1,6 @@
 package com.gmail.monajemi.am.android.qrscanner.fragment;
 
+import android.app.AlertDialog;
 import android.content.Intent;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
@@ -12,6 +13,7 @@ import android.os.Bundle;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import androidx.core.content.ContextCompat;
 import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.ItemTouchHelper;
 import androidx.recyclerview.widget.LinearLayoutManager;
@@ -28,12 +30,15 @@ import com.gmail.monajemi.am.android.qrscanner.Validation;
 import com.gmail.monajemi.am.android.qrscanner.adapter.HistoryAdapter;
 import com.gmail.monajemi.am.android.qrscanner.database.SqliteDatabase;
 import com.gmail.monajemi.am.android.qrscanner.model.History;
+import com.google.android.material.snackbar.BaseTransientBottomBar;
+import com.google.android.material.snackbar.Snackbar;
 
 import java.util.List;
 
 public class HistoryFragment extends Fragment implements HistoryAdapter.OnItemClickListener {
 
     private RecyclerView recyclerView;
+    private boolean undo = false;
 
     public HistoryFragment() {
     }
@@ -64,33 +69,52 @@ public class HistoryFragment extends Fragment implements HistoryAdapter.OnItemCl
 
             @Override
             public void onSwiped(@NonNull RecyclerView.ViewHolder viewHolder, int i) {
-                int position = viewHolder.getAdapterPosition();
-                History history = adapter.getItemAt(position);
-                database.deleteHistory(String.valueOf(history.getId()));
-                histories.remove(position);
-                adapter.notifyItemRemoved(position);
+                if (i == ItemTouchHelper.LEFT) {
+                    int position = viewHolder.getAdapterPosition();
+                    History history = adapter.getItemAt(position);
+                    histories.remove(position);
+                    adapter.notifyItemRemoved(position);
+                    Snackbar.make(recyclerView, getString(R.string.item_removed), BaseTransientBottomBar.LENGTH_SHORT)
+                            .setTextColor(ContextCompat.getColor(requireContext(),R.color.dark_grey))
+                            .setAction(getString(R.string.undo), view1 -> {
+                                undo = true;
+                                histories.add(position, history);
+                                adapter.notifyItemInserted(position);
+                            }).setActionTextColor(ContextCompat.getColor(requireContext(),R.color.white))
+                            .addCallback(new Snackbar.Callback() {
+                                @Override
+                                public void onDismissed(Snackbar snackbar, int event) {
+                                    if (!undo){
+                                        database.deleteHistory(String.valueOf(history.getId()));
+                                    }
+                                }
+
+                                @Override
+                                public void onShown(Snackbar snackbar) {
+                                }
+                            })
+                            .show();
+                }
             }
 
             @Override
             public void onChildDraw(@NonNull Canvas c, @NonNull RecyclerView recyclerView, @NonNull RecyclerView.ViewHolder viewHolder, float dX, float dY, int actionState, boolean isCurrentlyActive) {
                 super.onChildDraw(c, recyclerView, viewHolder, dX, dY, actionState, isCurrentlyActive);
-                Bitmap icon;
                 if (actionState == ItemTouchHelper.ACTION_STATE_SWIPE) {
-                    c.save();
+                    Bitmap icon;
                     View itemView = viewHolder.itemView;
                     float height = (float) itemView.getBottom() - (float) itemView.getTop();
                     float width = height / 3;
-                    if (dX <0) {
-                        RectF background = new RectF((float) itemView.getLeft(), (float) itemView.getTop(), dX, (float) itemView.getBottom());
-                        Paint p = new Paint();
+                    Paint p = new Paint();
+                    if (dX < 0) {
+                        RectF background = new RectF((float) itemView.getRight() + dX, (float) itemView.getTop(), (float) itemView.getRight(), (float) itemView.getBottom());
                         p.setColor(Color.BLACK);
                         c.drawRect(background, p);
-                        icon = BitmapFactory.decodeResource(getResources(), R.drawable.ic_baseline_delete);
-                        RectF icon_dest = new RectF((float) itemView.getLeft() + width, (float) itemView.getTop() + width, (float) itemView.getLeft() + 2 * width, (float) itemView.getBottom() - width);
+                        icon = BitmapFactory.decodeResource(requireActivity().getResources(), R.drawable.icons_remove);
+                        RectF icon_dest = new RectF((float) itemView.getRight() - 2 * width, (float) itemView.getTop() + width, (float) itemView.getRight() - width, (float) itemView.getBottom() - width);
                         c.drawBitmap(icon, null, icon_dest, p);
                     }
                 }
-                c.restore();
             }
         }).attachToRecyclerView(recyclerView);
         SnapHelper snapHelper = new LinearSnapHelper();
@@ -99,9 +123,15 @@ public class HistoryFragment extends Fragment implements HistoryAdapter.OnItemCl
 
     @Override
     public void onItemClick(String data) {
-        if (Validation.isValidURL(data)) {
-            Intent intent = new Intent(Intent.ACTION_VIEW, Uri.parse(data));
-            startActivity(intent);
-        }
+        AlertDialog.Builder alertDialog = new AlertDialog.Builder(getActivity()).setTitle(getString(R.string.scanned_data))
+                .setMessage(data).setNeutralButton(getString(R.string.ok), (dialogInterface, i) -> {
+
+                });
+        if (Validation.isValidURL(data))
+            alertDialog.setPositiveButton(getString(R.string.open_browser), (dialogInterface, i) -> {
+                Intent intent = new Intent(Intent.ACTION_VIEW, Uri.parse(data));
+                startActivity(intent);
+            });
+        alertDialog.show();
     }
 }
